+++ packages/core/src/index.ts
@@ -1,18 +1,22 @@
+import { Linker, MemoryLinker } from "./linkers";
 import { AutoDocsPlugin } from "./Plugin";
 import "./types";
 
 export * from "./Plugin";
-export * from "./analyzer";
+export type * from "./analyzer";
 export * from "./utils";
+export * from "./linkers";
 
 export class LambdaDocsBuilder<T extends AutoDocsTypes.AvailablePlugins> {
   public config: AutoDocsTypes.AutoDocsConfig<T>;
-  private _docs: Map<T, AutoDocsTypes.Plugins[T][]> = new Map();
+  private _docs: Linker<T>;
 
   private plugins: AutoDocsPlugin<T>[] = [];
 
   constructor(config: AutoDocsTypes.AutoDocsConfig<T>) {
     this.config = config;
+
+    this._docs = config.linker || new MemoryLinker();
     this.initPlugins();
   }
 
@@ -28,37 +32,57 @@ export class LambdaDocsBuilder<T extends AutoDocsTypes.AvailablePlugins> {
     });
   }
 
-  public async run() {
+  public async run<T extends AutoDocsTypes.AvailablePlugins>(): Promise<
+    Record<T, AutoDocsTypes.PluginResponse>
+  > {
+    const handlersFilter = await this._docs.pull();
+    const results: Record<
+      AutoDocsTypes.AvailablePlugins,
+      AutoDocsTypes.PluginResponse
+    > = {};
     this.plugins.forEach((plugin) => {
-      const handlersFilter = this._docs.get(plugin.type);
-
       if (handlersFilter) {
-        plugin.onBuild(handlersFilter, this);
+        const f = handlersFilter[plugin.type];
+        const map = f.map((item) => item.data);
+        results[plugin.type] = plugin.onBuild(map, this);
       }
     });
 
     this.plugins.forEach((plugin) => {
       plugin.onEnd(this);
     });
+
+    return results;
   }
 
-  public docs<T extends AutoDocsTypes.AvailablePlugins>(
+  public async docs<T extends AutoDocsTypes.AvailablePlugins>(
     type: T,
     docs: AutoDocsTypes.Plugins[T]
   ) {
-    if (
-      !this.plugins.some(
-        (plugin) => (plugin.type as AutoDocsTypes.AvailablePlugins) === type
-      )
-    ) {
+    const plugin = this.getPlugin(type);
+
+    if (!plugin) {
       throw new Error(`Plugin ${type} not found`);
     }
 
-    if (!this._docs.has(type)) {
-      this._docs.set(type, []);
-    }
-    this._docs.get(type)?.push(docs);
+    docs = plugin.onDoc(docs);
+
+    await this._docs.link({
+      data: docs,
+      plugin: type,
+      version:
+        "version" in docs ? (docs as { version: string }).version : "0.0.0",
+      description: "TODO",
+      name: "name" in docs ? (docs as { name: string }).name : "Unknown",
+    });
+
     return this;
+  }
+
+  private getPlugin(
+    type: AutoDocsTypes.AvailablePlugins
+  ): AutoDocsPlugin<T> | undefined {
+    return this.plugins.find((plugin) => plugin.type === type);
   }
 
   private isConcretePlugin(