+++ packages/plugins/other-plugin/src/index.ts
@@ -5,9 +5,4 @@ export class OtherApiDoc extends AutoDocsPlugin<"other"> {
   constructor() {
     super("other");
   }
-
-  onBuild(
-    docs: AutoDocsTypes.IOtherApi[],
-    builder: LambdaDocsBuilder<AutoDocsTypes.AvailablePlugins>
-  ): void {}
 }