+++ serverless-test/src/handlers/handler.ts
@@ -1,24 +1,44 @@
 import { otherImport } from "@/other/other-import";
+import { dynamicAutoDocs } from "@auto-docs/serverless-dynamic";
+import { LambdaDocsBuilder } from "@auto-docs/core";
+import { OpenApiDoc } from "@auto-docs/openapi-plugin";
+import * as Dynamic from "@auto-docs/serverless-dynamic";
+
 /**
  * @auto-docs
  * This is a test function. The endpoint is /hello.
  * @name Hello Endpoint
  * @version 1.1.2
  */
-export const hello = async (event: any) => {
+const helloBase = async (event: any) => {
   if (event.queryStringParameters && event.queryStringParameters.other) {
     return await otherImport();
   }
 
-  /* @auto-docs
-   * The return response is a User object or a message.
-   * @statusCode 201
-   * @schema { message: string }
-   */
   return {
-    message: "Go Serverless v1.0! Your function executed successfully!",
+    body: JSON.stringify({
+      hello: "World",
+    }),
+    description: "Wow this is so cool!",
+    schema: "com.drokt.HelloMessage",
+    statusCode: 200,
   };
 };
+
+const builder = new LambdaDocsBuilder({
+  name: "Test",
+  description: "Test",
+  pluginConfig: {
+    openApi: {
+      outputDir: "docs",
+      version: "1.0.0",
+    },
+  },
+  linker: new Dynamic.DynamoLinker("dynamicDocs"),
+  plugins: [OpenApiDoc],
+});
+
+export const hello = dynamicAutoDocs(helloBase, builder);
 
 /**
  * Bye!
@@ -42,3 +62,5 @@ export const bye = async (event: any) => {
     message: "Goodbye!",
   };
 };
+
+export const proxy = Dynamic.lambdaProxy();