+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @auto-docs/core
 
+## 1.1.0
+
+### Minor Changes
+
+- add dynamic handlers
+
 ## 1.0.1
 
 ### Patch Changes