+++ packages/plugins/other-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @auto-docs/other-plugin
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @auto-docs/core@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes