+++ packages/serverless-dynamic/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @auto-docs/openapi-plugin
+# @auto-docs/serverless
 
 ## 1.1.0
 
@@ -9,7 +9,20 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @auto-docs/openapi-plugin@1.1.0
   - @auto-docs/core@1.1.0
+
+## 1.1.1
+
+### Patch Changes
+
+- add custom config
+
+## 1.1.0
+
+### Minor Changes
+
+- added build command
 
 ## 1.0.1
 
@@ -17,4 +30,5 @@
 
 - add change set
 - Updated dependencies
+  - @auto-docs/openapi-plugin@1.0.1
   - @auto-docs/core@1.0.1