+++ packages/serverless/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @auto-docs/serverless
 
+## 1.2.0
+
+### Minor Changes
+
+- add dynamic handlers
+
+### Patch Changes
+
+- Updated dependencies
+  - @auto-docs/openapi-plugin@1.1.0
+  - @auto-docs/core@1.1.0
+
 ## 1.1.1
 
 ### Patch Changes